requirement.txt
chainlit
openai
python-dotenv
-e .
google-generativeai